use promise-chan for replacing callbacks
Originally we returned normal core.async channel and put! value on it
when api callback was called, and then immediately closed. Newly we
return promise-chan and don't close it, only in case of an error.

This has better semantics for callbacks. There can be multiple consumers
waiting for the same api-call result.

See https://dev.clojure.org/jira/browse/ASYNC-103

This should not require code rewrite because most of the time caller
is the only consumer and it does not pass go channel around. Newly this
allows greater flexibility.

Also newly we call set-last-error-args! to make args accessible in case
of errors.